fix: notify queries when executing group statements with result
