+++ sqldelight-compiler/integration-tests/src/test/kotlin/com/example/PlayerQueries.kt
@@ -252,6 +252,10 @@ public class PlayerQueries(
           |  FROM player
           |  WHERE player.rowid = last_insert_rowid()
           """.trimMargin(), mapper, 0)
+    } .also {
+      notifyQueries(781_651_682) { emit ->
+        emit("player")
+      }
     }
 
     override fun toString(): String = "Player.sq:insertAndReturn"
