+++ sqldelight-compiler/src/test/kotlin/app/cash/sqldelight/core/QueriesTypeTest.kt
@@ -876,7 +876,8 @@ class QueriesTypeTest {
 
     val dataQueries = File(result.outputDirectory, "com/example/DataQueries.kt")
     assertThat(result.compilerOutput).containsKey(dataQueries)
-    assertThat(result.compilerOutput[dataQueries].toString()).isEqualTo(
+    val queryString = result.compilerOutput[dataQueries].toString()
+    assertThat(queryString).isEqualTo(
       """
       |package com.example
       |
@@ -906,6 +907,10 @@ class QueriesTypeTest {
       |          |  VALUES (NULL)
       |          ""${'"'}.trimMargin(), 0)
       |      driver.executeQuery(${query.idForIndex(1).withUnderscores}, ""${'"'}SELECT last_insert_rowid()""${'"'}, mapper, 0)
+      |    } .also {
+      |      notifyQueries(${query.id.withUnderscores}) { emit ->
+      |        emit("data")
+      |      }
       |    }
       |
       |    override fun toString(): String = "Data.sq:insertAndReturn"
