+++ sqldelight-compiler/src/test/kotlin/app/cash/sqldelight/core/queries/async/AsyncSelectQueryTypeTest.kt
@@ -252,6 +252,10 @@ class AsyncSelectQueryTypeTest {
       |          |  FROM data
       |          |  WHERE id = last_insert_rowid()
       |          ""${'"'}.trimMargin(), mapper, 0).await()
+      |    } .also {
+      |      notifyQueries(${query.id.withUnderscores}) { emit ->
+      |        emit("data")
+      |      }
       |    }
       |  }
       |
